requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
ultralytics==8.0.196
python-dotenv==1.0.0
tenacity==8.2.3
kagglehub

